requirements.txt
streamlit==0.81.1
torch==1.7.0
torchvision==0.8.1
Pillow==8.2.0
